linux: kmd: static analysis TOP25 03/05 Part 9
Under the following path:
- drivers/media/platform/tegra/camera/fusa-capture/capture-vi-channel.c
- drivers/media/platform/tegra/camera/vi/mc_common.c
- drivers/platform/tegra/rtcpu/capture-ivc.c
- drivers/platform/tegra/rtcpu/ivc-bus.c
- drivers/platform/tegra/rtcpu/reset-group.c

Jira CAMERASW-32533

Reviewed-on: https://git-master.nvidia.com/r/c/linux-nv-oot/+/3314106
Tested-by: Patrick Young <payoung@nvidia.com>
GVS: buildbot_gerritrpt <buildbot_gerritrpt@nvidia.com>

+++ drivers/media/platform/tegra/camera/fusa-capture/capture-vi-channel.c
@@ -691,7 +691,7 @@ int vi_channel_drv_register(
 	mutex_unlock(&chdrv_lock);
 
 	for (i = 0; i < chan_drv->num_channels; i++) {
-		dev_t devt = MKDEV(vi_channel_major, i);
+		dev_t devt = MKDEV((unsigned long)vi_channel_major, i);
 
 		struct device *dev = &chan_drv->vi_capture_pdev->dev;
 		device_create(vi_channel_class, dev, devt, NULL,
